update: Scanフェーズの画面を別コンポーネントとして切り出し
