+++ homete/Resouces/Localizable.xcstrings
@@ -17,6 +17,9 @@
     "OK" : {
 
     },
+    "お手数ですが、再度デバイスを近づけて通信を行ってください" : {
+
+    },
     "キャンセル" : {
 
     },
@@ -83,6 +86,9 @@
 
     },
     "家事テンプレート" : {
+
+    },
+    "接続エラー" : {
 
     },
     "操作が完了しませんでした" : {
