+++ homete/Views/RegisterCohabitantView/SubViews/CompletedState/AnimationComponents/ConfettiRain/ConfettiRainPiece.swift
