+++ homete/Views/RegisterCohabitantView/SubViews/CompletedState/AnimationComponents/Cracker/CrackerRibbon.swift
