+++ homete/Views/RegisterCohabitantView/SubViews/CompletedState/AnimationComponents/Cracker/CrackerView.swift
