+++ homete/Views/RegisterCohabitantView/SubViews/ScanningState/CohabitantRegistrationInitialStateView.swift
